Restructure loader and loader tests code
Also applied `black` formatting to all loader-related files that were
involved in the restructuring.
