+++ tests/conftest.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 import yaml
 from httomo.runner.dataset import DataSet
-from httomo.runner.loader import DarksFlatsFileConfig
+from httomo.darks_flats import DarksFlatsFileConfig
 from httomo.ui_layer import _yaml_loader
