+++ tests/loaders/test_standard_tomo_loader.py
@@ -1,18 +1,20 @@
-import os
 from pathlib import Path
-from typing import Tuple, Union
+from typing import Tuple
 from unittest import mock
 
 import h5py
-from mpi4py import MPI
 import pytest
-from pytest_mock import MockerFixture
 import numpy as np
+from mpi4py import MPI
+from pytest_mock import MockerFixture
 
-from httomo.data.hdf.loaders import LoaderData
-from httomo.runner.dataset import DataSet
-from httomo.runner.loader import DarksFlatsFileConfig, Preview, PreviewConfig, PreviewDimConfig, RawAngles, StandardTomoLoader, UserDefinedAngles, get_darks_flats
-from ..testing_utils import make_mock_repo
+from httomo.darks_flats import DarksFlatsFileConfig
+from httomo.loaders.standard_tomo_loader import (
+    RawAngles,
+    StandardTomoLoader,
+    UserDefinedAngles,
+)
+from httomo.preview import PreviewConfig, PreviewDimConfig
 
 
 def make_standard_tomo_loader() -> StandardTomoLoader:
@@ -26,9 +28,7 @@ def make_standard_tomo_loader() -> StandardTomoLoader:
         data_path="/entry1/tomo_entry/data/data",
         image_key_path="/entry1/tomo_entry/instrument/detector/image_key",
     )
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry1/tomo_entry/data/rotation_angle"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     PREVIEW_CONFIG = PreviewConfig(
         angles=PreviewDimConfig(start=0, stop=180),
         detector_y=PreviewDimConfig(start=0, stop=128),
@@ -52,7 +52,7 @@ def make_standard_tomo_loader() -> StandardTomoLoader:
 
 def test_standard_tomo_loader_gives_h5py_dataset():
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = make_standard_tomo_loader()
@@ -62,7 +62,7 @@ def test_standard_tomo_loader_gives_h5py_dataset():
 
 def test_standard_tomo_loader_get_slicing_dim():
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = make_standard_tomo_loader()
@@ -107,18 +107,16 @@ def test_standard_tomo_loader_previewed_get_chunk_index_single_proc(
 ):
     IN_FILE_PATH = Path(__file__).parent.parent / "test_data/tomo_standard.nxs"
     DARKS_FLATS_CONFIG = DarksFlatsFileConfig(
-            file=IN_FILE_PATH,
-            data_path=standard_data_path,
-            image_key_path=standard_image_key_path,
-        )
-    ANGLES_CONFIG = RawAngles(
-            data_path="/entry1/tomo_entry/data/rotation_angle"
-            )
+        file=IN_FILE_PATH,
+        data_path=standard_data_path,
+        image_key_path=standard_image_key_path,
+    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     SLICING_DIM = 0
     COMM = MPI.COMM_WORLD
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -148,21 +146,27 @@ def test_standard_tomo_loader_previewed_get_chunk_index_single_proc(
                 angles=PreviewDimConfig(start=0, stop=180),
                 detector_y=PreviewDimConfig(start=0, stop=128),
                 detector_x=PreviewDimConfig(start=0, stop=160),
-            ), 0, 0,
+            ),
+            0,
+            0,
         ),
         (
             PreviewConfig(
                 angles=PreviewDimConfig(start=0, stop=180),
                 detector_y=PreviewDimConfig(start=5, stop=128),
                 detector_x=PreviewDimConfig(start=0, stop=160),
-            ), 1, 5,
+            ),
+            1,
+            5,
         ),
         (
             PreviewConfig(
                 angles=PreviewDimConfig(start=0, stop=180),
                 detector_y=PreviewDimConfig(start=0, stop=128),
                 detector_x=PreviewDimConfig(start=5, stop=160),
-            ), 2, 5,
+            ),
+            2,
+            5,
         ),
     ],
     ids=["no_cropping", "crop_det_y", "crop_det_x"],
@@ -181,25 +185,25 @@ def test_standard_tomo_loader_previewed_get_chunk_index_two_procs(
         data_path=standard_data_path,
         image_key_path=standard_image_key_path,
     )
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry1/tomo_entry/data/rotation_angle"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     SLICING_DIM = 0
     COMM = MPI.COMM_WORLD
 
     if crop_dim_index == 1:
         chunk_index = (
-            (0, crop_dim_start, 0) if COMM.rank == 0 else
-            (GLOBAL_DATA_SHAPE[0] // 2, crop_dim_start, 0)
+            (0, crop_dim_start, 0)
+            if COMM.rank == 0
+            else (GLOBAL_DATA_SHAPE[0] // 2, crop_dim_start, 0)
         )
     else:
         chunk_index = (
-            (0, 0, crop_dim_start) if COMM.rank == 0 else
-            (GLOBAL_DATA_SHAPE[0] // 2, 0, crop_dim_start)
+            (0, 0, crop_dim_start)
+            if COMM.rank == 0
+            else (GLOBAL_DATA_SHAPE[0] // 2, 0, crop_dim_start)
         )
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -259,14 +263,12 @@ def test_standard_tomo_loader_get_chunk_shape_single_proc(
         data_path=standard_data_path,
         image_key_path=standard_image_key_path,
     )
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry1/tomo_entry/data/rotation_angle"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     SLICING_DIM = 0
     COMM = MPI.COMM_WORLD
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -330,14 +332,12 @@ def test_standard_tomo_loader_get_chunk_shape_two_procs(
         data_path=standard_data_path,
         image_key_path=standard_image_key_path,
     )
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry1/tomo_entry/data/rotation_angle"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     SLICING_DIM = 0
     COMM = MPI.COMM_WORLD
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -403,9 +403,7 @@ def test_standard_tomo_loader_read_block_single_proc(
         data_path=standard_data_path,
         image_key_path=standard_image_key_path,
     )
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry1/tomo_entry/data/rotation_angle"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     SLICING_DIM = 0
     COMM = MPI.COMM_WORLD
 
@@ -427,7 +425,7 @@ def test_standard_tomo_loader_read_block_single_proc(
     expected_chunk_index = (BLOCK_START, 0, 0)
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -447,7 +445,7 @@ def test_standard_tomo_loader_read_block_single_proc(
     with h5py.File(IN_FILE_PATH, "r") as f:
         dataset: h5py.Dataset = f[standard_data_path]
         projs: np.ndarray = dataset[
-            PROJS_START + BLOCK_START: PROJS_START + BLOCK_START + BLOCK_LENGTH,
+            PROJS_START + BLOCK_START : PROJS_START + BLOCK_START + BLOCK_LENGTH,
             preview_config.detector_y.start : preview_config.detector_y.stop,
             preview_config.detector_x.start : preview_config.detector_x.stop,
         ]
@@ -510,9 +508,7 @@ def test_standard_tomo_loader_read_block_two_procs(
         data_path=standard_data_path,
         image_key_path=standard_image_key_path,
     )
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry1/tomo_entry/data/rotation_angle"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     SLICING_DIM = 0
     COMM = MPI.COMM_WORLD
 
@@ -525,7 +521,7 @@ def test_standard_tomo_loader_read_block_two_procs(
     )
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -543,8 +539,9 @@ def test_standard_tomo_loader_read_block_two_procs(
     block = loader.read_block(BLOCK_START, BLOCK_LENGTH)
 
     projs_start = (
-        preview_config.angles.start if COMM.rank == 0 else
-        (preview_config.angles.stop - preview_config.angles.start) // 2
+        preview_config.angles.start
+        if COMM.rank == 0
+        else (preview_config.angles.stop - preview_config.angles.start) // 2
     )
     # Index of block relative to the global data
     expected_global_index = (
@@ -558,7 +555,7 @@ def test_standard_tomo_loader_read_block_two_procs(
     with h5py.File(IN_FILE_PATH, "r") as f:
         dataset: h5py.Dataset = f[standard_data_path]
         projs: np.ndarray = dataset[
-            projs_start + BLOCK_START: projs_start + BLOCK_START + BLOCK_LENGTH,
+            projs_start + BLOCK_START : projs_start + BLOCK_START + BLOCK_LENGTH,
             preview_config.detector_y.start : preview_config.detector_y.stop,
             preview_config.detector_x.start : preview_config.detector_x.stop,
         ]
@@ -573,9 +570,7 @@ def test_standard_tomo_loader_read_block_adjust_for_darks_flats_single_proc():
     IN_FILE_PATH = Path(__file__).parent.parent / "test_data/k11_diad/k11-18014.nxs"
     DATA_PATH = "/entry/imaging/data"
     IMAGE_KEY_PATH = "/entry/instrument/imaging/image_key"
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry/imaging_sum/gts_theta_value"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry/imaging_sum/gts_theta_value")
     DARKS_CONFIG = FLATS_CONFIG = DarksFlatsFileConfig(
         file=IN_FILE_PATH,
         data_path=DATA_PATH,
@@ -590,7 +585,7 @@ def test_standard_tomo_loader_read_block_adjust_for_darks_flats_single_proc():
     COMM = MPI.COMM_WORLD
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -624,7 +619,7 @@ def test_standard_tomo_loader_read_block_adjust_for_darks_flats_single_proc():
     with h5py.File(IN_FILE_PATH, "r") as f:
         dataset: h5py.Dataset = f[DATA_PATH]
         projs: np.ndarray = dataset[
-            PROJS_START + BLOCK_START: PROJS_START + BLOCK_START + BLOCK_LENGTH
+            PROJS_START + BLOCK_START : PROJS_START + BLOCK_START + BLOCK_LENGTH
         ]
 
     assert block.global_index == expected_global_index
@@ -641,9 +636,7 @@ def test_standard_tomo_loader_read_block_adjust_for_darks_flats_two_procs():
     IN_FILE_PATH = Path(__file__).parent.parent / "test_data/k11_diad/k11-18014.nxs"
     DATA_PATH = "/entry/imaging/data"
     IMAGE_KEY_PATH = "/entry/instrument/imaging/image_key"
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry/imaging_sum/gts_theta_value"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry/imaging_sum/gts_theta_value")
     DARKS_CONFIG = FLATS_CONFIG = DarksFlatsFileConfig(
         file=IN_FILE_PATH,
         data_path=DATA_PATH,
@@ -658,7 +651,7 @@ def test_standard_tomo_loader_read_block_adjust_for_darks_flats_two_procs():
     COMM = MPI.COMM_WORLD
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -681,7 +674,7 @@ def test_standard_tomo_loader_read_block_adjust_for_darks_flats_two_procs():
     CHUNK_SHAPE = (
         GLOBAL_DATA_SHAPE[0] // 2,
         GLOBAL_DATA_SHAPE[1],
-        GLOBAL_DATA_SHAPE[2]
+        GLOBAL_DATA_SHAPE[2],
     )
 
     # Darks/flats are at indices 0 to 99 (and 3101 to 3200), projection data starts at index
@@ -700,7 +693,7 @@ def test_standard_tomo_loader_read_block_adjust_for_darks_flats_two_procs():
     with h5py.File(IN_FILE_PATH, "r") as f:
         dataset: h5py.Dataset = f[DATA_PATH]
         projs: np.ndarray = dataset[
-            projs_start + BLOCK_START: projs_start + BLOCK_START + BLOCK_LENGTH
+            projs_start + BLOCK_START : projs_start + BLOCK_START + BLOCK_LENGTH
         ]
 
     assert block.global_index == expected_global_index
@@ -716,7 +709,7 @@ def test_standard_tomo_loader_generates_block_with_angles():
     BLOCK_LENGTH = 2
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = make_standard_tomo_loader()
@@ -755,7 +748,7 @@ def test_standard_tomo_loader_user_defined_angles(
     )
 
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = StandardTomoLoader(
@@ -776,10 +769,9 @@ def test_standard_tomo_loader_user_defined_angles(
     np.testing.assert_array_equal(block.angles, EXPECTED_ANGLES)
 
 
-
 def test_standard_tomo_loader_closes_file(mocker: MockerFixture):
     with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
+        "httomo.darks_flats.get_darks_flats",
         return_value=(np.zeros(1), np.zeros(1)),
     ):
         loader = make_standard_tomo_loader()
@@ -793,9 +785,7 @@ def test_standard_tomo_loader_raises_error_slicing_dim(
     standard_data_darks_flats_config: DarksFlatsFileConfig,
 ):
     IN_FILE_PATH = Path(__file__).parent.parent / "test_data/tomo_standard.nxs"
-    ANGLES_CONFIG = RawAngles(
-        data_path="/entry1/tomo_entry/data/rotation_angle"
-    )
+    ANGLES_CONFIG = RawAngles(data_path="/entry1/tomo_entry/data/rotation_angle")
     PREVIEW_CONFIG = PreviewConfig(
         angles=PreviewDimConfig(start=0, stop=180),
         detector_y=PreviewDimConfig(start=0, stop=128),
@@ -804,10 +794,13 @@ def test_standard_tomo_loader_raises_error_slicing_dim(
     SLICING_DIM = 1
     COMM = MPI.COMM_WORLD
 
-    with mock.patch(
-        "httomo.runner.loader.get_darks_flats",
-        return_value=(np.zeros(1), np.zeros(1)),
-    ), pytest.raises(NotImplementedError):
+    with (
+        mock.patch(
+            "httomo.darks_flats.get_darks_flats",
+            return_value=(np.zeros(1), np.zeros(1)),
+        ),
+        pytest.raises(NotImplementedError),
+    ):
         _ = StandardTomoLoader(
             in_file=IN_FILE_PATH,
             data_path=standard_data_darks_flats_config.data_path,
@@ -819,331 +812,3 @@ def test_standard_tomo_loader_raises_error_slicing_dim(
             slicing_dim=SLICING_DIM,
             comm=COMM,
         )
-
-
-@pytest.mark.parametrize(
-    "preview_config, is_error_expected, err_str",
-    [
-        (
-            PreviewConfig(
-                angles=PreviewDimConfig(start=0, stop=221),
-                detector_y=PreviewDimConfig(start=0, stop=128),
-                detector_x=PreviewDimConfig(start=0, stop=160),
-            ),
-            True,
-            "Preview indices in angles dim exceed bounds of data: start=0, stop=221",
-        ),
-        (
-            PreviewConfig(
-                angles=PreviewDimConfig(start=0, stop=220),
-                detector_y=PreviewDimConfig(start=0, stop=129),
-                detector_x=PreviewDimConfig(start=0, stop=160),
-            ),
-            True,
-            "Preview indices in detector_y dim exceed bounds of data: start=0, stop=129",
-        ),
-        (
-            PreviewConfig(
-                angles=PreviewDimConfig(start=0, stop=220),
-                detector_y=PreviewDimConfig(start=0, stop=128),
-                detector_x=PreviewDimConfig(start=0, stop=161),
-            ),
-            True,
-            "Preview indices in detector_x dim exceed bounds of data: start=0, stop=161",
-        ),
-        (
-            PreviewConfig(
-                angles=PreviewDimConfig(start=220, stop=220),
-                detector_y=PreviewDimConfig(start=0, stop=128),
-                detector_x=PreviewDimConfig(start=0, stop=161),
-            ),
-            True,
-            "Start index in preview indices for angles dim is >= stop index: start=220, stop=220",
-        ),
-        (
-            PreviewConfig(
-                angles=PreviewDimConfig(start=0, stop=220),
-                detector_y=PreviewDimConfig(start=60, stop=50),
-                detector_x=PreviewDimConfig(start=0, stop=160),
-            ),
-            True,
-            "Start index in preview indices for detector_y dim is >= stop index: start=60, stop=50",
-        ),
-        (
-            PreviewConfig(
-                angles=PreviewDimConfig(start=0, stop=220),
-                detector_y=PreviewDimConfig(start=0, stop=128),
-                detector_x=PreviewDimConfig(start=50, stop=0),
-            ),
-            True,
-            "Start index in preview indices for detector_x dim is >= stop index: start=50, stop=0",
-        ),
-        (
-            PreviewConfig(
-                angles=PreviewDimConfig(start=0, stop=220),
-                detector_y=PreviewDimConfig(start=0, stop=128),
-                detector_x=PreviewDimConfig(start=0, stop=160),
-            ),
-            False,
-            "",
-        ),
-    ],
-    ids=[
-        "incorrect_angles_bounds",
-        "incorrect_det_y_bounds",
-        "incorrect_det_x_bounds",
-        "start_geq_stop_det_angles_bounds",
-        "start_geq_stop_det_y_bounds",
-        "start_geq_stop_det_x_bounds",
-        "all_correct_bounds",
-    ],
-)
-def test_preview_bound_checking(
-    standard_data_path: str,
-    standard_image_key_path: str,
-    preview_config: PreviewConfig,
-    is_error_expected: bool,
-    err_str: str,
-):
-    IN_FILE = Path(__file__).parent.parent / "test_data/tomo_standard.nxs"
-    f = h5py.File(IN_FILE, "r")
-    dataset = f[standard_data_path]
-    image_key = f[standard_image_key_path]
-
-    if is_error_expected:
-        with pytest.raises(ValueError, match=err_str):
-            _ = Preview(
-                preview_config=preview_config,
-                dataset=dataset,
-                image_key=image_key,
-            )
-    else:
-        preview = Preview(
-            preview_config=preview_config,
-            dataset=dataset,
-            image_key=image_key,
-        )
-        assert preview.config == preview_config
-
-    f.close()
-
-
-def test_preview_calculate_data_indices_excludes_darks_flats(
-    standard_data_path: str,
-    standard_image_key_path:str,
-):
-    IN_FILE_PATH = Path(__file__).parent.parent / "test_data/tomo_standard.nxs"
-    f = h5py.File(IN_FILE_PATH, "r")
-    dataset = f[standard_data_path]
-    image_key = f[standard_image_key_path]
-    all_indices: np.ndarray = image_key[:]
-    data_indices = np.where(all_indices == 0)[0]
-
-    config = PreviewConfig(
-        angles=PreviewDimConfig(start=0, stop=220),
-        detector_y=PreviewDimConfig(start=0, stop=128),
-        detector_x=PreviewDimConfig(start=0, stop=160),
-    )
-    preview = Preview(
-        preview_config=config,
-        dataset=dataset,
-        image_key=image_key,
-    )
-    assert not np.array_equal(preview.data_indices, all_indices)
-    assert np.array_equal(preview.data_indices, data_indices)
-    assert preview.config.angles == PreviewDimConfig(start=0, stop=180)
-    f.close()
-
-
-def test_preview_with_no_image_key():
-    IN_FILE_PATH = (
-        Path(__file__).parent.parent /
-            "test_data/i12/separate_flats_darks/i12_dynamic_start_stop180.nxs"
-    )
-    f = h5py.File(IN_FILE_PATH, "r")
-    DATA_PATH = "1-TempPlugin-tomo/data"
-    ANGLES, DET_Y, DET_X = (724, 10, 192)
-    expected_indices = list(range(ANGLES))
-    config = PreviewConfig(
-        angles=PreviewDimConfig(start=0, stop=ANGLES),
-        detector_y=PreviewDimConfig(start=0, stop=DET_Y),
-        detector_x=PreviewDimConfig(start=0, stop=DET_X),
-    )
-    preview = Preview(
-        preview_config=config,
-        dataset=f[DATA_PATH],
-        image_key=None,
-    )
-    assert np.array_equal(preview.data_indices, expected_indices)
-
-
-@pytest.mark.parametrize(
-    "previewed_shape",
-    [(100, 128, 160), (180, 10, 160), (180, 128, 10)],
-    ids=["crop_angles_dim", "crop_det_y_dim", "crop_det_x_dim"],
-)
-def test_preview_global_shape(
-    standard_data_path: str,
-    standard_image_key_path:str,
-    previewed_shape: Tuple[int, int, int],
-):
-    IN_FILE_PATH = Path(__file__).parent.parent / "test_data/tomo_standard.nxs"
-    f = h5py.File(IN_FILE_PATH, "r")
-    dataset = f[standard_data_path]
-    image_key = f[standard_image_key_path]
-
-    config = PreviewConfig(
-        angles=PreviewDimConfig(start=0, stop=previewed_shape[0]),
-        detector_y=PreviewDimConfig(start=0, stop=previewed_shape[1]),
-        detector_x=PreviewDimConfig(start=0, stop=previewed_shape[2]),
-    )
-    preview = Preview(
-        preview_config=config,
-        dataset=dataset,
-        image_key=image_key,
-    )
-    assert preview.global_shape == previewed_shape
-
-
-@pytest.mark.parametrize(
-    "preview_config",
-    [
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=128),
-            detector_x=PreviewDimConfig(start=0, stop=160),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=10),
-            detector_x=PreviewDimConfig(start=0, stop=160),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=10, stop=20),
-            detector_x=PreviewDimConfig(start=0, stop=160),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=128),
-            detector_x=PreviewDimConfig(start=0, stop=10),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=128),
-            detector_x=PreviewDimConfig(start=10, stop=20),
-        ),
-    ],
-    ids=[
-        "no_cropping",
-        "crop_det_y_start_0",
-        "crop_det_y_start_10",
-        "crop_det_x_start_0",
-        "crop_det_x_start_10",
-    ],
-)
-def test_get_darks_flats_same_file_same_dataset(
-    standard_data_path: str,
-    standard_data_darks_flats_config: DarksFlatsFileConfig,
-    preview_config: PreviewConfig,
-):
-    IN_FILE_PATH = Path(__file__).parent.parent / "test_data/tomo_standard.nxs"
-
-    loaded_darks, loaded_flats = get_darks_flats(
-        standard_data_darks_flats_config,
-        standard_data_darks_flats_config,
-        preview_config,
-    )
-
-    FLATS_START = 180
-    FLATS_END = 199
-    DARKS_START = 200
-    DARKS_END = 219
-    with h5py.File(IN_FILE_PATH, "r") as f:
-        dataset: h5py.Dataset = f[standard_data_path]
-        flats = dataset[
-            FLATS_START : FLATS_END + 1,
-            preview_config.detector_y.start : preview_config.detector_y.stop,
-            preview_config.detector_x.start : preview_config.detector_x.stop,
-        ]
-        darks = dataset[
-            DARKS_START : DARKS_END + 1,
-            preview_config.detector_y.start : preview_config.detector_y.stop,
-            preview_config.detector_x.start : preview_config.detector_x.stop,
-        ]
-
-    np.testing.assert_array_equal(loaded_flats, flats)
-    np.testing.assert_array_equal(loaded_darks, darks)
-
-
-@pytest.mark.parametrize(
-    "preview_config",
-    [
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=128),
-            detector_x=PreviewDimConfig(start=0, stop=160),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=10),
-            detector_x=PreviewDimConfig(start=0, stop=160),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=10, stop=20),
-            detector_x=PreviewDimConfig(start=0, stop=160),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=128),
-            detector_x=PreviewDimConfig(start=0, stop=10),
-        ),
-        PreviewConfig(
-            angles=PreviewDimConfig(start=0, stop=180),
-            detector_y=PreviewDimConfig(start=0, stop=128),
-            detector_x=PreviewDimConfig(start=10, stop=20),
-        ),
-    ],
-    ids=[
-        "no_cropping",
-        "crop_det_y_start_0",
-        "crop_det_y_start_10",
-        "crop_det_x_start_0",
-        "crop_det_x_start_10",
-    ],
-)
-def test_get_darks_flats_different_file(preview_config: PreviewConfig):
-    DARKS_CONFIG = DarksFlatsFileConfig(
-        file=Path(__file__).parent.parent / "test_data/i12/separate_flats_darks/dark_field.h5",
-        data_path="/1-NoProcessPlugin-tomo/data",
-        image_key_path=None,
-    )
-    FLATS_CONFIG = DarksFlatsFileConfig(
-        file=Path(__file__).parent.parent / "test_data/i12/separate_flats_darks/flat_field.h5",
-        data_path="/1-NoProcessPlugin-tomo/data",
-        image_key_path=None,
-    )
-
-    loaded_darks, loaded_flats = get_darks_flats(
-        DARKS_CONFIG,
-        FLATS_CONFIG,
-        preview_config,
-    )
-
-    with h5py.File(DARKS_CONFIG.file, "r") as f:
-        darks = f[DARKS_CONFIG.data_path][
-            :,
-            preview_config.detector_y.start : preview_config.detector_y.stop,
-            preview_config.detector_x.start : preview_config.detector_x.stop,
-        ]
-
-    with h5py.File(FLATS_CONFIG.file, "r") as f:
-        flats = f[FLATS_CONFIG.data_path][
-            :,
-            preview_config.detector_y.start : preview_config.detector_y.stop,
-            preview_config.detector_x.start : preview_config.detector_x.stop,
-        ]
-
-    np.testing.assert_array_equal(loaded_flats, flats)
-    np.testing.assert_array_equal(loaded_darks, darks)
